backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
numpy>=1.26.0
pydantic>=2.5.0
python-dotenv>=1.0.0
fpdf2>=2.7.0